refactor CFlow implementation
