+++ anomalib/models/cflow/__init__.py
@@ -1,7 +1,4 @@
-"""Real-Time  Unsupervised Anomaly Detection via Conditional Normalizing Flows.
-
-[CW-AD](https://arxiv.org/pdf/2107.12571v1.pdf)
-"""
+"""Real-Time  Unsupervised Anomaly Detection via Conditional Normalizing Flows."""
 
 # Copyright (C) 2020 Intel Corporation
 #
@@ -16,3 +13,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions
 # and limitations under the License.
+
+from .model import CflowModel
+
+__all__ = ["CflowModel"]
